+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.4.0] - 2020-11-26
+### Added
+- Campaigns
+  - add contact to campaign
+  - remove contact from campaign
+  - list contacts campaigns from Contact model
+### Changed
+- .rubocop (normal indentation)
 ## [2.3.11] - 2020-09-30
 ### Added
 - permit `referer` in submission